+++ view/src/pages/FlashCardsPage/FlashCardsPage.tsx
@@ -1,12 +1,14 @@
 import { Box, Stack } from '@chakra-ui/react';
 import SocialLinks from '../../components/SocialLinks';
 import NavBar from '../../components/NavBar';
+import CardTemplate from '../../components/CardTemplate';
 
 const FlashCardsPage = () => {
     return (
         <>
             <Stack minHeight="100vh">
                 <NavBar />
+                <CardTemplate />
                 <Box marginTop="auto" marginBottom="5%">
                     <SocialLinks />
                 </Box>
